// src/app/features/notification/types/model/notification-update-model.ts
import {NotificationTypesEnum} from '../common/notification-types.enum';

export interface INotificationUpdateModel {
  isNotifyTeachers: boolean;
  isNotifyAdmins: boolean;
  adminEmails: Array<string>;
  notifyType: NotificationTypesEnum;
  notifyDay: number;
  notifyTime: string;
  notifyBeforeDays: number;
  attestationYearsPeriod: number;
  requiredInternshipHours: number;
}